PdfToHtml/backend/requirements.txt
fastapi
uvicorn
httpx 
requests
openai
python-dotenv
pydantic